+++ pybmix/core/pybmixcpp/CMakeLists.txt
@@ -151,7 +151,7 @@ endif ()
 add_library(bayesmix OBJECT)
 target_sources(bayesmix PUBLIC ${PROTO_SRCS} ${PROTO_HDRS})
 add_subdirectory(bayesmix/src)
-add_subdirectory(py_hier)
+add_subdirectory(python_embedding)
 target_include_directories(bayesmix PUBLIC ${INCLUDE_PATHS})
 target_link_libraries(bayesmix PUBLIC ${LINK_LIBRARIES})
 target_compile_options(bayesmix PUBLIC ${COMPILE_OPTIONS})
