+++ pybmix/core/pybmixcpp/python_embedding/auxiliary_functions.cc
