+++ pybmix/core/pybmixcpp/python_embedding/python_hierarchy.cc
@@ -98,7 +98,7 @@ void PythonHierarchy::set_hypers_from_proto(
 }
 
 void PythonHierarchy::set_module(const std::string &module_name) {
-    std::cout << "Using implementation: " << module_name << std::endl;
+    std::cout << "Using hierarchy implementation in " << module_name << ".py" << std::endl;
 
     hier_implementation = py::module_::import(module_name.c_str());
